+++ pidman/pid/models.py
@@ -54,7 +54,7 @@ class Policy(models.Model):
     '''
     created_at = models.DateTimeField("Date Created", auto_now_add=True)
     commitment = models.TextField("Commitment Statement")
-    title      = models.CharField(unique=True, max_length=255)
+    title = models.CharField(unique=True, max_length=255)
 
     objects = PolicyManager()
 
@@ -95,6 +95,9 @@ class Domain(MPTTModel):
     def __unicode__(self):
         return self.name
 
+    def __repr__(self):
+        return '<Domain %s>' % unicode(self)
+
     def get_policy(self):
         '''Return the policy for this domain or from parent domain if inherited.
         Raises a :class:`DomainException` if no policy is found.
@@ -110,20 +113,23 @@ class Domain(MPTTModel):
         else:
             raise DomainException("No Policy Exists")
 
+    def show_policy(self):
+        try:
+            return self.get_policy()
+        except DomainException:
+            return 'No policy set'
+    show_policy.name = 'Policy'
+
     def subdomain_count(self):
         return self.get_descendant_count()
-    subdomain_count.short_description = '# subdomains'
+    subdomain_count.short_description = '# collections'
 
     def num_pids(self):
         '''Number of :class:`Pid` instances in this domain, including those in
         any subdomains.'''
         domain_ids = self.get_descendants(include_self=True).values_list('id', flat=True)
         return Pid.objects.filter(domain__in=domain_ids).count()
-        pid_count = self.pid_set.count()
-        for subdomain in self.collections.all():
-            pid_count += subdomain.pid_set.count()
-        return pid_count
-    num_pids.short_description = "Total pids"
+    num_pids.short_description = "# pids"
 
 class PidManager(models.Manager):
     def get_by_natural_key(self, noid):